refactor: 패키지, 접근자
