+++ src/main/java/com/nextstep/javaRacing/car/MoveStrategy.java
@@ -1,4 +1,4 @@
-package com.nextstep.javaRacing;
+package com.nextstep.javaRacing.car;
 
 public interface MoveStrategy {
