+++ src/main/java/com/nextstep/javaRacing/car/RandomMoveStrategy.java
@@ -1,4 +1,4 @@
-package com.nextstep.javaRacing;
+package com.nextstep.javaRacing.car;
 
 import java.util.Random;
