+++ src/test/java/com/nextstep/javaRacing/car/RaceCarTest.java
@@ -1,4 +1,4 @@
-package com.nextstep.javaRacing;
+package com.nextstep.javaRacing.car;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
