+++ src/test/java/com/nextstep/javaRacing/race/RaceFactoryTest.java
@@ -1,4 +1,4 @@
-package com.nextstep.javaRacing;
+package com.nextstep.javaRacing.race;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
