+++ src/test/java/com/nextstep/javaRacing/race/RaceTest.java
@@ -1,5 +1,8 @@
-package com.nextstep.javaRacing;
+package com.nextstep.javaRacing.race;
 
+import com.nextstep.javaRacing.car.Car;
+import com.nextstep.javaRacing.car.MoveStrategy;
+import com.nextstep.javaRacing.car.RaceCar;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
